Refactor/shared config decks in a different module
